api: move ICR to dedicated fixture file
