+++ carbonmark-api/test/routes/vintages/get.test.ts
@@ -2,8 +2,9 @@ import { FastifyInstance } from "fastify";
 import nock from "nock";
 import { aProject } from "../../../src/.generated/mocks/marketplace.mocks";
 import { GRAPH_URLS, ICR_API } from "../../../src/app.constants";
+import { COUNTRY_CODES, VINTAGES } from "../../../test/fixtures/icr";
 import { build } from "../../helper";
-import { COUNTRY_CODES, DEV_URL, ERROR, VINTAGES } from "../../test.constants";
+import { DEV_URL, ERROR } from "../../test.constants";
 
 describe("GET /vintages", () => {
   let fastify: FastifyInstance;
